requirements.txt
django == 5.1.3
django-mathfilters == 1.0.0
langchain == 0.3.5
langchain-groq == 0.11.0
humanize == 4.11.0
python == 3.10
tensorflow == 2.10
pillow == 10.4.0
numpy == 1.26.4